AS3 fighting game engine
The Birth Engine-- named after the game it's being developed for, Under
Mcleod In-Birth-- is a high performance fighting game engine developed
in FlashDevelop and ActionScript 3. It is still currently in
development.  It will feature smooth, blitted animations and graphics at
60FPS, 3v3 Tag-Team support (inspired by Marvel vs Capcom 2-3), hitstun
support, custom events, combos and combo damage scaling, etc.